refactor: Phase 1.1 - consolidate catalog files into unified module
Consolidate 4 catalog files into single catalog/ module:
- catalog.py (818 LOC) → deleted
- catalog_service.py (273 LOC) → deleted
- service_layer/catalog.py (85 LOC) → deleted
- models/catalog.py (45 LOC) → moved to catalog/models.py

New structure:
- catalog/service.py (582 LOC) - unified implementation
- catalog/models.py (45 LOC) - data models
- catalog/__init__.py - public API

Changes:
- Merged CatalogBuilder, DatabaseDiscoveryService, SchemaMetadataCollector
- Merged build_catalog() function and CatalogService wrapper
- Fixed export_sql_from_catalog signature to match original
- Updated all imports (6 files):
  - commands/discover.py
  - pipelines/catalog.py
  - service_layer/__init__.py
  - models/__init__.py
  - __init__.py

Results:
- LOC: 14,947 → 14,456 (-491 LOC, 3.3% reduction)
- Files: 4 catalog files → 2 catalog files
- Zero duplication in catalog module
- All functionality preserved

Refs: v1.8.0 Phase 1.1
