+++ src/snowcli_tools/service_layer/__init__.py
@@ -1,6 +1,6 @@
 """Service layer abstractions for snowcli-tools."""
 
-from .catalog import CatalogService
+from ..catalog import CatalogService
 from .dependency import DependencyService
 from .query import QueryService
